Modularize env/val
