+++ src/native/str.rs
@@ -1,4 +1,4 @@
-use crate::eval::{Environment};
+use crate::environment::Environment;
 
 pub fn add_str_procs(_env: &mut Environment) {
 	// TODO
